regex_app/requirements.txt
django==5.1
djangorestframework==3.15.2
pandas==2.2.2
openai==0.28.0
django-cors-headers==4.4.0
xlrd>=2.0.1  # For reading .xls files
openpyxl>=3.0.9  # For reading .xlsx files

python-dotenv==1.0.1